requirements.txt
requests
python-dotenv
sentence-transformers
faiss-cpu
numpy
pyyaml
javalang
crewai
gradio